pro_game/wsgi.py
import os
from decouple import config
from django.core.wsgi import get_wsgi_application

os.environ.setdefault(
    "DJANGO_SETTINGS_MODULE",
    config("DJANGO_SETTINGS_MODULE", default="pro_game.settings.settings"),
)

application = get_wsgi_application()
